// SalesViewPCF/cds-generated/enums/axa_departmentfulfillmentstatus_axa_departmentfulfillmentstatus_axa_fulfillmentstatus.ts
/* eslint-disable*/
// Enum axa_departmentfulfillmentstatus_axa_departmentfulfillmentstatus_axa_fulfillmentstatus
export const enum axa_departmentfulfillmentstatus_axa_departmentfulfillmentstatus_axa_fulfillmentstatus {
  Done = 752580000,
  WIP = 752580001,
  Notstarted = 752580002,
  Cancelled = 752580003,
}
  